Moving the JsonMovie class from the Repository to a separate class
