+++ app/src/main/java/ru/androidacademy/gooseeye/api/responses/MoviesResponse.kt
@@ -3,7 +3,7 @@ package ru.androidacademy.gooseeye.api.responses
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import ru.androidacademy.gooseeye.data.Repository.JsonMovie
+import ru.androidacademy.gooseeye.data.models.JsonMovie
 
 @ExperimentalSerializationApi
 @Serializable
